Add hook to review an updated position
